adds ses as email provider
related #40
