+++ package.json
@@ -45,6 +45,7 @@
   },
   "homepage": "https://github.com/childline-yoti/online-content-reporter#readme",
   "dependencies": {
+    "aws-sdk": "^2.53.0",
     "axios": "^0.15.3",
     "babel-core": "^6.24.0",
     "babel-loader": "^6.4.1",
